+++ crates/guest-rust/src/lib.rs
@@ -895,7 +895,7 @@ pub mod rt {
 
 #[cfg(feature = "async")]
 pub use wit_bindgen_rt::async_support::{
-    backpressure_set, block_on, spawn, AbiBuffer, FutureRead, FutureReader, FutureWrite,
-    FutureWriteCancel, FutureWriteError, FutureWriter, StreamRead, StreamReader, StreamResult,
-    StreamWrite, StreamWriter,
+    backpressure_set, block_on, spawn, yield_async, yield_blocking, AbiBuffer, FutureRead,
+    FutureReader, FutureWrite, FutureWriteCancel, FutureWriteError, FutureWriter, StreamRead,
+    StreamReader, StreamResult, StreamWrite, StreamWriter,
 };
