+++ tests/runtime-async/async/cancel-import/runner.rs
@@ -3,8 +3,8 @@ include!(env!("BINDINGS"));
 use crate::my::test::i::*;
 use futures::task::noop_waker_ref;
 use std::future::Future;
-use std::pin::Pin;
-use std::task::{Context, Poll};
+use std::task::Context;
+use wit_bindgen::yield_async;
 
 fn main() {
     println!("test cancelling an import in progress");
@@ -79,7 +79,7 @@ fn main() {
         // Let the subtask's completion notification make its way to our task
         // here.
         for _ in 0..5 {
-            yield_().await;
+            yield_async().await;
         }
 
         // Now cancel the import, despite it actually being done. This should
@@ -114,7 +114,7 @@ fn main() {
         // notification that it's entered the "STARTED" state.
         backpressure_set(false);
         for _ in 0..5 {
-            yield_().await;
+            yield_async().await;
         }
 
         // Now cancel the `starting_import`. This should correctly pick up the
@@ -130,27 +130,4 @@ fn main() {
         tx1.write(()).await.unwrap();
         started_import.await;
     });
-}
-
-async fn yield_() {
-    #[derive(Default)]
-    struct Yield {
-        yielded: bool,
-    }
-
-    impl Future for Yield {
-        type Output = ();
-
-        fn poll(mut self: Pin<&mut Self>, context: &mut Context<'_>) -> Poll<()> {
-            if self.yielded {
-                Poll::Ready(())
-            } else {
-                self.yielded = true;
-                context.waker().wake_by_ref();
-                Poll::Pending
-            }
-        }
-    }
-
-    Yield::default().await;
 }
